requirements.txt
Flask==2.3.4
Flask-SQLAlchemy==3.1.3
python-dotenv==1.0.0
resend==1.0.0
